Add `atmos aws` commands
